login - authentication principal added

+++ src/main/java/com/ythwork/soda/web/AccountController.java
@@ -10,6 +10,7 @@ import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.IanaLinkRelations;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,12 +19,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.ythwork.soda.domain.Account;
+import com.ythwork.soda.domain.Member;
 import com.ythwork.soda.dto.AccountAddInfo;
 import com.ythwork.soda.dto.AccountInfo;
 import com.ythwork.soda.exception.AccountAlreadyExistsException;
 import com.ythwork.soda.exception.AccountNotFoundException;
 import com.ythwork.soda.exception.EntityAlreadyExistsException;
 import com.ythwork.soda.exception.EntityNotFoundException;
+import com.ythwork.soda.exception.NotAllowedMemberException;
 import com.ythwork.soda.hateoas.AccountModelAssembler;
 import com.ythwork.soda.service.AccountService;
 
@@ -39,13 +43,14 @@ public class AccountController {
 	}
 	
 	@GetMapping
-	public CollectionModel<EntityModel<AccountInfo>> allAccounts() {
-		List<EntityModel<AccountInfo>> accounts = accountService.getAllAccountInfos().stream()
+	public CollectionModel<EntityModel<AccountInfo>> allAccounts(@AuthenticationPrincipal Member member) {
+		List<AccountInfo> accounts = accountService.getAllAccountInfosByMember(member);
+		List<EntityModel<AccountInfo>> entityModels = accounts.stream()
 				.map(assembler::toModel)
 				.collect(Collectors.toList());
 		
-		return CollectionModel.of(accounts, 
-				linkTo(methodOn(AccountController.class).allAccounts()).withSelfRel());
+		return CollectionModel.of(entityModels, 
+				linkTo(methodOn(AccountController.class).allAccounts(member)).withSelfRel());
 	}
 	
 	@GetMapping("/{id}")
@@ -60,7 +65,11 @@ public class AccountController {
 	}
 	
 	@PostMapping
-	ResponseEntity<?> addAccount(@RequestBody AccountAddInfo accountAddInfo) {
+	ResponseEntity<?> addAccount(@RequestBody AccountAddInfo accountAddInfo, @AuthenticationPrincipal Member member) {
+		if(accountAddInfo.getMemberId() != member.getId()) {
+			throw new NotAllowedMemberException("회원의 계좌 번호가 아닙니다.");
+		}
+		
 		AccountInfo accountInfo = null;
 		try {
 			accountInfo = accountService.addAccountToBoard(accountAddInfo);
@@ -76,7 +85,18 @@ public class AccountController {
 	}
 	
 	@DeleteMapping("/{id}")
-	ResponseEntity<?> deleteAccount(@PathVariable Long id) {
+	ResponseEntity<?> deleteAccount(@PathVariable Long id, @AuthenticationPrincipal Member member) {
+		Account account = null;
+		try {
+			account = accountService.findById(id);
+		} catch(EntityNotFoundException e) {
+			throw new AccountNotFoundException(e.getMessage(), e);
+		}
+		
+		if(account.getMember().getId() != member.getId()) {
+			throw new NotAllowedMemberException("회원의 계좌 번호가 아닙니다. 삭제할 수 없습니다.");
+		}
+		
 		accountService.deleteAccount(id);
 		
 		return ResponseEntity.noContent().build();
